requirements.txt
easyocr
pillow
watchdog
pyaudio
numpy
pyautogui
opencv-python
scikit-image
flask
flask-socketio
flask-cors
langchain-community
langchain-core
tiktoken
chromadb
